Advanced

Advanced features

Objective function

BlockDecomposition.objectivevaluemagnitude — Function.

objectivevaluemagnitude(m::JuMP.Model, magnitude)

Set the magnitude of the objective function of the model m to magnitude

source

BlockDecomposition.objectivevaluelowerbound — Function.

objectivevaluelowerbound(m::JuMP.Model, lb)

Set the lower bound of the objective function of the model m to lb

source

BlockDecomposition.objectivevalueupperbound — Function.

objectivevalueupperbound(m::JuMP.Model, ub)

Set the upper bound of the objective function of the model m to ub

source

Branching priorities

BlockDecomposition.branchingpriorityinmaster — Function.

branchingpriorityinmaster(x::JuMP.JuMPContainer, subproblem::Tuple{Symbol, Union{Tuple, Integer}}, priority)

Assign to the variables x defined in the subproblem subproblem the priority value priority in master.

branchingpriorityinmaster(x, (:B_SP, 1), 2)

The variable x defined in the Benders subproblem with id 1 will have the branching priority value 2in the master.

source

BlockDecomposition.branchingpriorityinsubproblem — Function.

branchingpriorityinsubproblem(x::JuMP.JuMPContainer, subproblem::Tuple{Symbol, Union{Tuple, Integer}}, priority)

Assign to the variables x defined in the subproblem subproblem the priority value priority in subproblems.

branchingpriorityinsubproblem(x, (:B_SP, 1), 2)

The variable x defined in the Benders subproblem with id 1 will have the branching priority value 2in subproblems.

source